realtor-v0/requirements-production.txt
# Production Requirements for Propertism Realty Advisors LLP
# SCCB-47 Compliance

# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0

# Database
psycopg2-binary>=2.9.6  # PostgreSQL adapter

# Translation
django-modeltranslation>=0.18.0

# Static Files & Media
whitenoise>=6.5.0  # Static file serving
Pillow>=10.0.0  # Image processing

# Production Server
gunicorn>=21.2.0  # WSGI server

# Environment Variables
python-dotenv>=1.0.0

# Security & Monitoring
django-redis>=5.3.0  # Redis cache (optional)

# Email (choose one)
# sendgrid>=6.10.0  # If using SendGrid
# mailgun>=0.1.1    # If using Mailgun

# Testing
colorama>=0.4.6  # For colored test output

# Optional but Recommended
# sentry-sdk>=1.32.0  # Error tracking
# django-storages>=1.13.0  # S3/Cloud storage
# boto3>=1.28.0  # AWS SDK (if using S3)
